api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
scipy>=1.11.0
pytest>=7.4.0

# Additional dependencies for production
python-multipart>=0.0.6
cryptography>=41.0.0